% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{multi}
\alias{multi}
\alias{multi,ANY,ANY-method}
\alias{multi,numeric,ANY-method}
\alias{multi,numeric,integer-method}
\title{An s4 generic}
\usage{
multi(x, y)

\S4method{multi}{numeric,ANY}(x, y)

\S4method{multi}{numeric,integer}(x, y)
}
\description{
An s4 generic
}
